Migrated to using futures and tested non-blocking code.
